+++ src/main/java/com/avab/avab/apiPayload/exception/S3Exception.java
@@ -1,9 +1,9 @@
-package com.avab.avab.apiPayload.exception.S3;
+package com.avab.avab.apiPayload.exception;
 
 import com.avab.avab.apiPayload.code.BaseErrorCode;
-import com.avab.avab.apiPayload.exception.GeneralException;
 
 public class S3Exception extends GeneralException {
+
     public S3Exception(BaseErrorCode code) {
         super(code);
     }
